Fix root x/y to be FP3232 on a barrier event

+++ include/X11/extensions/XInput2.h
@@ -343,8 +343,8 @@ typedef struct {
     int           sourceid;
     Window        window;
     Window        root;
-    int           root_x;
-    int           root_y;
+    double        root_x;
+    double        root_y;
     double        dx;
     double        dy;
     double        raw_dx;
